backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
python-dotenv==1.0.0
pydantic==2.10.1
loguru==0.7.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
numpy==1.24.3
scipy==1.11.1
pydub==0.25.1
azure-storage-blob==12.18.3
cloudinary==1.36.0
